Changed launch file and IMU reader node
A lot of other files are also changed, most of them are however probably ignorable. This code is from 18/05/2018 and was developed by Lucas van Zuylen. It includes a workaround for a bug where the SDK would not read the firmware version of the Matrice 100 and therefore could not launch.